// src/sdks/systemV2ApiDocs.ts
export type getSystemLibUnitPageProps = {
  pageNum: number;
  pageSize: number;
  unit?: number | null | undefined;
  unitClass?: number | null | undefined;
};
// ...
export interface OperLog {
  companyId?: number;

  content?: string;

  createBy?: number;

  createTime?: Date;

  id?: number;

  moduleType?:
    | '10'
    | '11'
    | '12'
    | '13'
    | '14'
    | '15'
    | '16'
    | '17'
    | '20'
    | '30'
    | '31'
    | '32'
    | '33'
    | '34'
    | '35'
    | '36'
    | '37';

  operType?: '1' | '2' | '3';

  realName?: string;
  updateBy?: number;
  updateTime?: Date;

  username?: string;
}